+++ pwndbg/aglib/__init__.py
@@ -9,14 +9,6 @@ regs = None
 def load_aglib():
     import pwndbg.aglib.argv
     import pwndbg.aglib.ctypes
-    import pwndbg.aglib.disasm
-    import pwndbg.aglib.disasm.aarch64
-    import pwndbg.aglib.disasm.arm
-    import pwndbg.aglib.disasm.mips
-    import pwndbg.aglib.disasm.ppc
-    import pwndbg.aglib.disasm.riscv
-    import pwndbg.aglib.disasm.sparc
-    import pwndbg.aglib.disasm.x86
     import pwndbg.aglib.dynamic
     import pwndbg.aglib.elf
     import pwndbg.aglib.file
