+++ pwndbg/arguments.py
@@ -11,8 +11,8 @@ from typing import Tuple
 from capstone import CS_GRP_INT
 
 import pwndbg.aglib.arch
-import pwndbg.aglib.disasm
 import pwndbg.aglib.disasm.arch
+import pwndbg.aglib.disassembly
 import pwndbg.aglib.file
 import pwndbg.aglib.memory
 import pwndbg.aglib.proc
