+++ pwndbg/commands/branch.py
@@ -5,8 +5,7 @@ import argparse
 import gdb
 from capstone import CS_GRP_JUMP
 
-import pwndbg.aglib.disasm
-import pwndbg.aglib.disasm.arch
+import pwndbg.aglib.disassembly
 import pwndbg.color.message as message
 import pwndbg.commands
 import pwndbg.gdblib.bpoint
@@ -28,7 +27,7 @@ class BreakOnConditionalBranch(pwndbg.gdblib.bpoint.Breakpoint):
         # We need to re-run the enhancement process, since now the PC == instruction.address,
         # where previously it was not. The enhancement process will figure out if all the conditions
         # this branch requires in order to be taken have been met.
-        assistant = pwndbg.aglib.disasm.arch.get_disassembly_assistant_for_current_arch()
+        assistant = pwndbg.aglib.disassembly.get_disassembly_assistant_for_current_arch()
         assistant.enhance(self.instruction)
         condition_met = self.instruction.is_conditional_jump_taken
 
@@ -92,7 +91,7 @@ def install_breakpoint(branch, taken: bool) -> None:
             return
 
     # We should've picked something by now, or errored out.
-    instruction = pwndbg.aglib.disasm.one(address)
+    instruction = pwndbg.aglib.disassembly.one(address)
     if instruction is None:
         print(message.error(f"Could not decode instruction at address {address:#x}"))
         return
@@ -105,7 +104,7 @@ def install_breakpoint(branch, taken: bool) -> None:
         return
 
     # Not all architectures have assistants we can use for conditionals.
-    if not pwndbg.aglib.disasm.arch.arch_has_disassembly_assistant():
+    if not pwndbg.aglib.disassembly.arch_has_disassembly_assistant():
         print(
             message.error(
                 "The current architecture is not supported for breaking on conditional branches"
