+++ pwndbg/enhance.py
@@ -14,7 +14,7 @@ from typing import Tuple
 
 import pwndbg
 import pwndbg.aglib.arch
-import pwndbg.aglib.disasm
+import pwndbg.aglib.disassembly
 import pwndbg.aglib.memory
 import pwndbg.aglib.strings
 import pwndbg.aglib.typeinfo
@@ -112,7 +112,7 @@ def enhance(
         rwx = exe = False
 
     if exe:
-        pwndbg_instr = pwndbg.aglib.disasm.one(value)
+        pwndbg_instr = pwndbg.aglib.disassembly.one(value)
         if pwndbg_instr:
             # For telescoping, we don't want the extra spaces between the mnemonic and operands
             # which are baked in during enhancement. This removes those spaces.
